move ui toolkit stuff into src/ui
